+++ internal/balancer/ctx.go
@@ -8,20 +8,13 @@ type (
 
 type Endpoint interface {
 	NodeID() uint32
+
+	// Choose calls from balancer if this endpoint to request chosen or not
+	Choose(chosen bool)
 }
 
 func WithEndpoint(ctx context.Context, endpoint Endpoint) context.Context {
 	return context.WithValue(ctx, ctxEndpointKey{}, endpoint)
-}
-
-type nodeID uint32
-
-func (n nodeID) NodeID() uint32 {
-	return uint32(n)
-}
-
-func WithNodeID(ctx context.Context, id uint32) context.Context {
-	return WithEndpoint(ctx, nodeID(id))
 }
 
 func ContextEndpoint(ctx context.Context) (e Endpoint, ok bool) {