+++ internal/table/session.go
@@ -53,6 +53,15 @@ type session struct {
 	closeOnce sync.Once
 }
 
+func (s *session) Choose(chosen bool) {
+	if s == nil {
+		return
+	}
+	if !chosen {
+		s.SetStatus(options.SessionClosing)
+	}
+}
+
 func (s *session) NodeID() uint32 {
 	if s == nil {
 		return 0
@@ -165,7 +174,7 @@ func (s *session) Close(ctx context.Context) (err error) {
 		onDone := trace.TableOnSessionDelete(s.config.Trace(), &ctx, s)
 
 		_, err = s.tableService.DeleteSession(
-			balancer.WithEndpoint(ctx, s),
+			ctx,
 			&Ydb_Table.DeleteSessionRequest{
 				SessionId: s.id,
 				OperationParams: operation.Params(ctx,